add GPU creating tree

+++ inc/init.h
@@ -1,6 +1,6 @@
 #ifndef INIT_H
 #define INIT_H
 
-void init( double *x, double *y, double *mass, const int n, const double region, const double boxsize, const double maxmass);
+void init( double *x, double *y, double *mass, const unsigned long n, const double region, const double boxsize, const double maxmass);
 
 #endif
